Revert "feat: split main.rs"
