--- src/lib.rs
@@ -1,5 +1,0 @@
-pub mod auth;
-pub mod task;
-
-pub use auth::*;
-pub use task::*;
